refactor(app): move outbound ports under dedicated module
