+++ src/app/cmd/browse/metadata.rs
@@ -197,9 +197,9 @@ mod tests {
     use std::time::Instant;
 
     use crate::app::model::app_state::AppState;
-    use crate::app::ports::connection_store::MockConnectionStore;
-    use crate::app::ports::metadata::MockMetadataProvider;
-    use crate::app::ports::query_executor::MockQueryExecutor;
+    use crate::app::ports::outbound::connection_store::MockConnectionStore;
+    use crate::app::ports::outbound::metadata::MockMetadataProvider;
+    use crate::app::ports::outbound::query_executor::MockQueryExecutor;
     use crate::app::ports::{DbOperationError, RenderOutput, RenderResult, Renderer};
     use crate::app::services::AppServices;
     use crate::app::update::action::Action;
