+++ src/app/cmd/connection.rs
@@ -215,9 +215,9 @@ mod tests {
     use crate::app::cmd::effect::Effect;
     use crate::app::cmd::test_support::*;
     use crate::app::model::app_state::AppState;
-    use crate::app::ports::connection_store::MockConnectionStore;
-    use crate::app::ports::metadata::MockMetadataProvider;
-    use crate::app::ports::query_executor::MockQueryExecutor;
+    use crate::app::ports::outbound::connection_store::MockConnectionStore;
+    use crate::app::ports::outbound::metadata::MockMetadataProvider;
+    use crate::app::ports::outbound::query_executor::MockQueryExecutor;
     use crate::app::ports::{
         ConnectionStoreError, DsnBuilder, RenderOutput, RenderResult, Renderer,
     };
