+++ src/app/cmd/runner.rs
@@ -403,9 +403,9 @@ impl EffectRunner {
 mod tests {
     use super::*;
     use crate::app::cmd::test_support::*;
-    use crate::app::ports::connection_store::MockConnectionStore;
-    use crate::app::ports::metadata::MockMetadataProvider;
-    use crate::app::ports::query_executor::MockQueryExecutor;
+    use crate::app::ports::outbound::connection_store::MockConnectionStore;
+    use crate::app::ports::outbound::metadata::MockMetadataProvider;
+    use crate::app::ports::outbound::query_executor::MockQueryExecutor;
     use crate::app::ports::{RenderOutput, RenderResult};
     use crate::app::services::AppServices;
     use crate::domain::{DatabaseMetadata, TableSummary};
