+++ src/app/ports/mod.rs
@@ -7,24 +7,6 @@
 pub mod inbound;
 pub mod outbound;
 
-pub mod clipboard;
-pub mod config_writer;
-pub mod connection_store;
-pub mod db_capabilities;
-pub mod db_operation_error;
-pub mod ddl_generator;
-pub mod dsn_builder;
-pub mod er_exporter;
-pub mod er_log_writer;
-pub mod folder_opener;
-pub mod graphviz;
-pub mod metadata;
-pub mod query_executor;
-pub mod query_history;
-pub mod renderer;
-pub mod service_file;
-pub mod sql_dialect;
-
 pub use inbound::{handle_input, InputEvent, InputKeyCombo, Key, Modifiers};
 pub use outbound::{
     ClipboardError,
