+++ src/app/ports/outbound/clipboard.rs
