+++ src/app/ports/outbound/connection_store.rs
