+++ src/app/ports/outbound/db_capabilities.rs
