+++ src/app/ports/outbound/db_operation_error.rs
