+++ src/app/ports/outbound/er_log_writer.rs
