+++ src/app/ports/outbound/folder_opener.rs
