+++ src/app/ports/outbound/mod.rs
@@ -1,17 +1,35 @@
-pub use super::clipboard::{ClipboardError, ClipboardWriter};
-pub use super::config_writer::{ConfigWriter, ConfigWriterError};
-pub use super::connection_store::{ConnectionStore, ConnectionStoreError};
-pub use super::db_capabilities::{DatabaseCapabilities, DatabaseCapabilityProvider, InspectorFeature};
-pub use super::db_operation_error::DbOperationError;
-pub use super::ddl_generator::DdlGenerator;
-pub use super::dsn_builder::DsnBuilder;
-pub use super::er_exporter::{ErDiagramExporter, ErExportError, ErExportResult};
-pub use super::er_log_writer::ErLogWriter;
-pub use super::folder_opener::{FolderOpenError, FolderOpener};
-pub use super::graphviz::{GraphvizError, GraphvizRunner, ViewerError, ViewerLauncher};
-pub use super::metadata::MetadataProvider;
-pub use super::query_executor::QueryExecutor;
-pub use super::query_history::{QueryHistoryError, QueryHistoryStore};
-pub use super::renderer::{RenderError, RenderOutput, RenderResult, Renderer};
-pub use super::service_file::{PgServiceEntryReader, ServiceFileError};
-pub use super::sql_dialect::SqlDialect;
+pub mod clipboard;
+pub mod config_writer;
+pub mod connection_store;
+pub mod db_capabilities;
+pub mod db_operation_error;
+pub mod ddl_generator;
+pub mod dsn_builder;
+pub mod er_exporter;
+pub mod er_log_writer;
+pub mod folder_opener;
+pub mod graphviz;
+pub mod metadata;
+pub mod query_executor;
+pub mod query_history;
+pub mod renderer;
+pub mod service_file;
+pub mod sql_dialect;
+
+pub use self::clipboard::{ClipboardError, ClipboardWriter};
+pub use self::config_writer::{ConfigWriter, ConfigWriterError};
+pub use self::connection_store::{ConnectionStore, ConnectionStoreError};
+pub use self::db_capabilities::{DatabaseCapabilities, DatabaseCapabilityProvider, InspectorFeature};
+pub use self::db_operation_error::DbOperationError;
+pub use self::ddl_generator::DdlGenerator;
+pub use self::dsn_builder::DsnBuilder;
+pub use self::er_exporter::{ErDiagramExporter, ErExportError, ErExportResult};
+pub use self::er_log_writer::ErLogWriter;
+pub use self::folder_opener::{FolderOpenError, FolderOpener};
+pub use self::graphviz::{GraphvizError, GraphvizRunner, ViewerError, ViewerLauncher};
+pub use self::metadata::MetadataProvider;
+pub use self::query_executor::QueryExecutor;
+pub use self::query_history::{QueryHistoryError, QueryHistoryStore};
+pub use self::renderer::{RenderError, RenderOutput, RenderResult, Renderer};
+pub use self::service_file::{PgServiceEntryReader, ServiceFileError};
+pub use self::sql_dialect::SqlDialect;
