+++ src/app/ports/outbound/query_executor.rs
