+++ src/app/ports/outbound/query_history.rs
