+++ src/app/update/action.rs
@@ -6,10 +6,10 @@ use crate::app::model::shared::key_sequence::Prefix;
 use crate::app::model::sql_editor::completion::CompletionCandidate;
 use crate::app::policy::write::write_guardrails::WritePreview;
 use crate::app::ports::DbOperationError;
-use crate::app::ports::clipboard::ClipboardError;
-use crate::app::ports::connection_store::ConnectionStoreError;
-use crate::app::ports::folder_opener::FolderOpenError;
-use crate::app::ports::query_history::QueryHistoryError;
+use crate::app::ports::outbound::clipboard::ClipboardError;
+use crate::app::ports::outbound::connection_store::ConnectionStoreError;
+use crate::app::ports::outbound::folder_opener::FolderOpenError;
+use crate::app::ports::outbound::query_history::QueryHistoryError;
 use crate::domain::connection::{ConnectionNameError, ConnectionProfile, ServiceEntry};
 use std::collections::HashMap;
